miden_project

Enum DependencyResolutionError

pub enum DependencyResolutionError {
    NoSolution(Box<DerivationTree<PackageId, VersionSet, String>>),
    FailedRetreivingDependencies {
        package: PackageId,
        version: Version,
        error: String,
    },
    FailedToChooseVersion {
        package: PackageId,
        error: String,
    },
    Cancelled {
        reason: String,
    },
}

The error type raised when PackageResolver::resolve fails.

Variants

NoSolution(Box<DerivationTree<PackageId, VersionSet, String>>)

No solution was possible due to conflicts or unsatisified constraints.

FailedRetreivingDependencies

Resolution could not proceed because one or more dependent packages were unavailable.

Fields

package: PackageId

version: Version

error: String

FailedToChooseVersion

Resolution could not proceed because the resolver was unable to choose an appropriate version for a given package.

Fields

package: PackageId

error: String

Cancelled

Resolution was cancelled by the implementation of [pubgrub::DependencyProvider::should_cancel] - typically due to timeout or some other external signal.

Fields

reason: String