Crate miden_project

Re-exports

pub use miden_assembly_syntax::semver;

Modules

ast

This module and its children define the abstract syntax tree representation of the miden-project.toml file and its variants (i.e. workspace-level vs package-level).

Structs

Dependency

Represents a project/package dependency declaration

LexicographicWord

A Word wrapper with lexicographic ordering.

Package

The representation of an individual package in a Miden project

PackageId

A type that represents the unique identifier for packages in a super::PackageIndex.

PackageIndex

PackageIndex is used as an in-memory database of package information needed for dependency resolution for Miden projects. It records and can provide information about:

PackageResolver

As implied by the name, this is the implementation of the dependency/package resolver.

Profile

Represents configuration options for a specific build profile, e.g. release

SemVer

SemVer version as defined by https://semver.org.

Target

Represents build target configuration

Uri

A URI reference that specifies the location of a source file, whether on disk, on the network, or elsewhere.

Version

The representation of versioning information associated with packages in the package index.

VersionReq

SemVer version requirement describing the intersection of some version comparators, such as >=1.2.3, <1.8.

VersionSet

This type is an implementation detail of the dependency resolver provided by super::PackageIndex.

Word

A unit of data consisting of 4 field elements.

Workspace

Represents a Miden project workspace.

Enums

DependencyResolutionError

The error type raised when PackageResolver::resolve fails.

DependencyVersionScheme

Represents the versioning requirement and resolution method for a specific dependency.

GitRevision

A reference to a revision in Git

InvalidDependencySpecError

Linkage

This represents the way in which a dependent will link against a dependency during assembly.

TargetType

The type of a specific target provided by the current project

Value

Representation of a TOML value.

VersionRequirement

Represents a requirement on a specific version (or versions) of a dependency.

Type Aliases

Map

An alias for alloc::collections::BTreeMap.

Metadata

Represents arbitrary metadata in key/value format

MetadataSet

Represents a set of named metadata tables, where each table is represented by Metadata.