miden_project

Module ast

This module and its children define the abstract syntax tree representation of the miden-project.toml file and its variants (i.e. workspace-level vs package-level).

The AST is used for parsing and rendering the TOML representation, but after validation and resolution of inherited properties, the AST is translated to a simpler structure that does not need to represent the complexity of the on-disk format.

Structs

BinTarget

DependencySpec

Represents information about a project dependency needed to resolve it to a Miden package

LibTarget

PackageConfig

Package configuration which can be defined at both the workspace and package level

PackageDetail

Package properties which may be inherited from a parent workspace

Profile

Represents configuration options for a specific build profile, e.g. release

ProjectFile

Represents the miden-project.toml structure of an individual package

WorkspaceFile

Represents a workspace-level miden-project.toml file

Enums

MidenProject

Represents all possible variants of miden-project.toml